Revert "Customize plugin icon"
