+++ src/main/kotlin/ai/devchat/core/DevChatClient.kt
@@ -1,12 +1,14 @@
 package ai.devchat.core
 
-import ai.devchat.common.HttpClient
 import ai.devchat.common.Log
 import ai.devchat.common.PathUtils
 import ai.devchat.plugin.localServicePort
-import ai.devchat.storage.CONFIG
 import kotlinx.coroutines.*
+import kotlinx.coroutines.channels.awaitClose
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.catch
+import kotlinx.coroutines.flow.flowOn
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
@@ -14,11 +16,16 @@ import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.serializer
-import java.net.URL
+import okhttp3.*
+import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.RequestBody.Companion.toRequestBody
+import java.io.IOException
 import java.nio.file.Files
 import java.nio.file.Paths
 import java.time.Instant
 import kotlin.system.measureTimeMillis
+
 
 
 inline fun <reified T> T.asMap(): Map<String, Any?> where T : @Serializable Any {
@@ -202,13 +209,10 @@ fun timeThis(block: suspend () -> Unit) {
     }
 }
 
-
 class DevChatClient {
     private val scope: CoroutineScope = CoroutineScope(Dispatchers.IO)
     private val baseURL get() =  "http://localhost:$localServicePort"
     private var job: Job? = null
-
-    val client: HttpClient = HttpClient()
 
     companion object {
         const val DEFAULT_LOG_MAX_COUNT = 10000
@@ -216,8 +220,93 @@ class DevChatClient {
         const val RETRY_INTERVAL: Long = 500  // ms
         const val MAX_RETRIES: Int = 10
     }
+    private val client = OkHttpClient()
     private val json = Json { ignoreUnknownKeys = true }
 
+    private inline fun <reified T> get(
+        path: String,
+        queryParams: Map<String, Any?> = emptyMap()
+    ): T? {
+        Log.info("GET request to [$baseURL$path] with request parameters: $queryParams")
+        val urlBuilder = "$baseURL$path".toHttpUrlOrNull()?.newBuilder() ?: return null
+        queryParams.forEach { (k, v) -> urlBuilder.addQueryParameter(k, v.toString()) }
+        val url = urlBuilder.build()
+        val request = Request.Builder().url(url).get().build()
+        var retries = MAX_RETRIES
+        while (retries > 0) {
+            try {
+                client.newCall(request).execute().use { response ->
+                    if (!response.isSuccessful) throw IOException(
+                        "Unsuccessful response: ${response.code} ${response.message}"
+                    )
+                    val result = response.body?.string()?.let {
+                        json.decodeFromString<T>(it)
+                    }
+                    return result
+                }
+            } catch (e: IOException) {
+                Log.warn("$e, retrying...")
+                retries--
+                Thread.sleep(RETRY_INTERVAL)
+            } catch (e: Exception) {
+                Log.warn(e.toString())
+                return null
+            }
+        }
+        return null
+    }
+
+    private inline fun <reified T, reified R> post(path: String, body: T? = null): R? {
+        Log.info("POST request to [$baseURL$path] with request body: $body")
+        val url = "$baseURL$path".toHttpUrlOrNull() ?: return null
+        val requestBody = json.encodeToString(serializer(), body).toRequestBody("application/json".toMediaType())
+        val request = Request.Builder().url(url).post(requestBody).build()
+        var retries = MAX_RETRIES
+        while (retries > 0) {
+            try {
+                val response: Response = client.newCall(request).execute()
+                if (!response.isSuccessful) throw IOException(
+                    "Unsuccessful response: ${response.code} ${response.message}"
+                )
+                val result = response.body?.let {
+                    json.decodeFromString<R>(it.string())
+                }
+                return result
+            } catch (e: IOException) {
+                Log.warn("$e, retrying...")
+                retries--
+                Thread.sleep(RETRY_INTERVAL)
+            } catch (e: Exception) {
+                Log.warn(e.toString())
+                return null
+            }
+        }
+        return null
+    }
+
+    private inline fun <reified T, reified R> streamPost(path: String, body: T? = null): Flow<R> = callbackFlow<R> {
+        Log.info("POST request to [$baseURL$path] with request body: $body")
+        val url = "$baseURL$path".toHttpUrlOrNull() ?: return@callbackFlow
+        val requestJson = json.encodeToString(serializer(), body)
+        val requestBody = requestJson.toRequestBody("application/json".toMediaType())
+        val request = Request.Builder().url(url).post(requestBody).build()
+        val call = client.newCall(request)
+
+        val response = call.execute()
+        if (!response.isSuccessful) {
+            throw IOException("Unexpected code $response")
+        }
+        response.body?.byteStream()?.use {inputStream ->
+            val buffer = ByteArray(8192) // 8KB buffer
+            var bytesRead: Int
+            while (inputStream.read(buffer).also { bytesRead = it } != -1) {
+                val chunk = buffer.copyOf(bytesRead).toString(Charsets.UTF_8)
+                send(json.decodeFromString(chunk))
+            }
+        }
+        close()
+        awaitClose { call.cancel() }
+    }.flowOn(Dispatchers.IO)
 
     fun message(
         message: ChatRequest,
@@ -227,7 +316,7 @@ class DevChatClient {
     ) {
         cancelMessage()
         job = scope.launch {
-            client.streamPost<ChatRequest, ChatResponse>("$baseURL/message/msg", message)
+            streamPost<ChatRequest, ChatResponse>("/message/msg", message)
                 .catch { e ->
                     onError(e.toString())
                     Log.warn("Error on sending message: $e")
@@ -246,13 +335,13 @@ class DevChatClient {
     }
 
     fun getWorkflowList(): List<Workflow>? {
-        return client.get("$baseURL/workflows/list")
+        return get("/workflows/list")
     }
     fun getWorkflowConfig(): WorkflowConfig? {
-        return client.get("$baseURL/workflows/config")
+        return get("/workflows/config")
     }
     fun updateWorkflows() {
-        val response: UpdateWorkflowResponse? = client.post<String?, _>("$baseURL/workflows/update")
+        val response: UpdateWorkflowResponse? = post<String?, _>("/workflows/update")
         Log.info("Update workflows response: $response")
     }
 
@@ -264,7 +353,7 @@ class DevChatClient {
         } else {
             body["filepath"] = PathUtils.createTempFile(jsonData, "devchat_log_insert_", ".json")
         }
-        val response: LogInsertRes? = client.post("$baseURL/logs/insert", body)
+        val response: LogInsertRes? = post("/logs/insert", body)
         if (body.containsKey("filepath")) {
             try {
                 Files.delete(Paths.get(body["filepath"]!!))
@@ -275,13 +364,13 @@ class DevChatClient {
         return response
     }
     fun deleteLog(logHash: String): LogDeleteRes? {
-         return client.post("$baseURL/logs/delete", mapOf(
+         return post("/logs/delete", mapOf(
             "workspace" to PathUtils.workspace,
             "hash" to logHash
         ))
     }
     fun getTopicLogs(topicRootHash: String, offset: Int = 0, limit: Int = DEFAULT_LOG_MAX_COUNT): List<ShortLog> {
-        return client.get<List<ShortLog>>("$baseURL/topics/$topicRootHash/logs", mapOf(
+        return get<List<ShortLog>>("/topics/$topicRootHash/logs", mapOf(
                 "limit" to limit,
                 "offset" to offset,
                 "workspace" to PathUtils.workspace,
@@ -293,54 +382,15 @@ class DevChatClient {
             "offset" to offset,
             "workspace" to PathUtils.workspace,
         )
-        return client.get<List<Topic>?>("$baseURL/topics", queryParams).orEmpty()
+        return get<List<Topic>?>("/topics", queryParams).orEmpty()
     }
 
     fun deleteTopic(topicRootHash: String) {
-        val response: Map<String, String>? = client.post("$baseURL/topics/delete", mapOf(
+        val response: Map<String, String>? = post("/topics/delete", mapOf(
             "topic_hash" to topicRootHash,
             "workspace" to PathUtils.workspace,
         ))
         Log.info("deleteTopic response data: $response")
-    }
-
-    fun getWebappUrl(): String? {
-        val apiBase = CONFIG["providers.devchat.api_base"] as String
-        val urlOrPath: String? = client.get("$apiBase/addresses/webapp")
-        if (urlOrPath.isNullOrEmpty()) {
-            Log.warn("No webapp url found");
-            return null
-        }
-        var href = ""
-        href = if (urlOrPath.startsWith("http://") || urlOrPath.startsWith("https://")) {
-            urlOrPath
-        } else {
-            URL(URL(apiBase), urlOrPath).toString()
-        }
-        if (href.endsWith('/')) {
-            href = href.dropLast(1)
-        }
-        if (href.endsWith("/api")) {
-            href = href.dropLast(4)
-        }
-        Log.info("Webapp url: $href")
-        return href
-    }
-
-    fun getIconUrl(): String {
-        try {
-            val webappUrl = getWebappUrl()
-            if (!webappUrl.isNullOrEmpty()) {
-                val iconsUrl = URL(URL(webappUrl), "/api/v1/plugin/icons/")
-                val res: Map<String, String?>? = client.get(URL(iconsUrl, "filename/intellij").toString())
-                res?.get("filename")?.let {
-                    return URL(iconsUrl, it).toString()
-                }
-            }
-        } catch (e: Exception) {
-            Log.warn(e.toString())
-        }
-        return "/icons/pluginIcon_dark.svg"
     }
 
     private fun cancelMessage() {
