+++ src/main/kotlin/ai/devchat/plugin/hints/ChatCVProviderBase.kt
@@ -1,13 +1,10 @@
 package ai.devchat.plugin.hints
 
-import ai.devchat.core.DC_CLIENT
 import ai.devchat.core.DevChatActions
 import ai.devchat.core.handlers.SendUserMessageHandler
 import ai.devchat.plugin.DevChatToolWindow
 import com.alibaba.fastjson.JSONObject
-import com.intellij.codeInsight.codeVision.CodeVisionAnchorKind
-import com.intellij.codeInsight.codeVision.CodeVisionEntry
-import com.intellij.codeInsight.codeVision.CodeVisionRelativeOrdering
+import com.intellij.codeInsight.codeVision.*
 import com.intellij.codeInsight.codeVision.ui.model.ClickableTextCodeVisionEntry
 import com.intellij.codeInsight.hints.codeVision.CodeVisionProviderBase
 import com.intellij.codeInsight.hints.settings.language.isInlaySettingsEditor
@@ -16,10 +13,7 @@ import com.intellij.openapi.roots.ProjectFileIndex
 import com.intellij.openapi.util.IconLoader
 import com.intellij.openapi.util.TextRange
 import com.intellij.openapi.wm.ToolWindowManager
-import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiFile
-import com.intellij.psi.SmartPointerManager
-import com.intellij.psi.SyntaxTraverser
+import com.intellij.psi.*
 import com.intellij.psi.util.elementType
 import java.awt.event.MouseEvent
 import javax.swing.Icon
@@ -28,7 +22,7 @@ abstract class ChatCVProviderBase : CodeVisionProviderBase() {
 
     abstract fun buildPayload(editor: Editor, element: PsiElement): JSONObject
     open fun getIcon(): Icon? {
-        return IconLoader.getIcon(DC_CLIENT.getIconUrl(), this::class.java.classLoader)
+        return IconLoader.getIcon("/icons/pluginIcon_dark.svg", this::class.java.classLoader)
     }
 
     override fun computeForEditor(editor: Editor, file: PsiFile): List<Pair<TextRange, CodeVisionEntry>> {
